Add files via upload
A higher education institute is looking for the Upgraded network for its two Campuses from which one is in the Melbourne and other is in the Sydney. Also, each campus has two different sectors that is School of IT and Business. It is required to access either campus from one place in the new design therefore, the new network should deploy the admittance of each campus to the admin and staff. Additionally, there need to be some of configurations that will able to provide essential services to the users. Another aspect are restrictions and configuration in the network. To integrate all these features, institute’s network should be upgraded into two separate Local Area Networks. These LANs are able to categorised different buildings such as laboratories and other offices which are associated to the School of IT and School of Business. Both Campuses will be connected with the help of WAN and can be monitored by Admin. 

The project scope is the essential aspect which helps to accomplish project success. In this report, there is need to develop a typical network design of a higher education institute which has further sections separated as Melbourne and Sydney campuses. Both campuses need to be connected with each other through WAN. The administrator should have access to the whole network. Moreover, both networks should be able to access remotely with the help of Network Monitoring System. In short, the scopes are elaborated as follow:
•	Enterprise two Campuses network design (Separate LAN): It requires the redesigning of LAN, VLAN, Data centre (DC) and Quality of service (QoS).
•	Enhance enterprise edge availability or remote access: It can be done by reforming WAN and it will provide communication between campuses.
•	Secured network should be provided in both campuses: Settings done in a way that students will not be able to use other resources of Campus through library.
•	Internet, printing services and wireless access availability in the certain sections of campus.